retinanet/requirements.txt
keras==2.1.2
tensorflow-gpu==1.3.0
pandas
opencv-python
jupyter
git+https://github.com/broadinstitute/keras-resnet
git+https://github.com/vickyliin/keras-retinanet.git
h5py
scikit-image
tqdm
nose
git+https://github.com/tensorflow/models.git#egg=tensorflow-models-research&subdirectory=research
